create media POST,PUT,DELETE

+++ portfolio/src/main/java/com/portfolio/dao/ToolsAndTechnologiesDaoJpa.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.portfolio.entity.ToolsAndTechnologies;
 
-public interface ToolsAndTechnologiesDaoJpa extends JpaRepository<ToolsAndTechnologies, Long> {
+public interface ToolsAndTechnologiesDaoJpa extends JpaRepository<ToolsAndTechnologies, Integer> {
  
 	ToolsAndTechnologies findById(int id);
 }
